+++ libs/hapify/generate-config/src/hpf-generate-config.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/node
+#!/usr/bin/env node
 import { getHapifyOptions } from './generate-config';
 
 (async () => getHapifyOptions())().catch((err) => console.error(err));