+++ libs/schematics/src/generators/hapify-library/helpers/normalize-options.helper.ts
@@ -3,7 +3,10 @@ import { relative } from 'path';
 import { getWorkspaceLayout, TargetConfiguration, Tree } from '@nrwl/devkit';
 import * as deepmerge from 'deepmerge';
 
-import { getNormalizedProjectDefaultsOptions } from '../../../helpers';
+import {
+  getImportPrefixPath,
+  getNormalizedProjectDefaultsOptions,
+} from '../../../helpers';
 import {
   DEFAULT_IMPORT_REPLACEMENTS,
   DEFAULT_SECONDARY_ENTRY_POINTS,
@@ -53,14 +56,24 @@ export function normalizeOptions(
     options.hapifyModelsJson,
   );
 
+  const importPrefixPath = getImportPrefixPath(tree, directory);
+
   // Format import replacement
-  const hapifyImportReplacements = [
-    ...new Set(
-      hapifyTemplates.flatMap(
-        (template) => DEFAULT_IMPORT_REPLACEMENTS[template],
-      ),
-    ),
-  ];
+  const hapifyImportReplacements: Record<string, string> = hapifyTemplates
+    .flatMap((template) => DEFAULT_IMPORT_REPLACEMENTS[template])
+    .map((template) => [template, importPrefixPath + template])
+    .concat(
+      hapifyTemplates.length === 1
+        ? [['mock', `${importPrefixPath}${hapifyTemplates[0]}/mock`]]
+        : [],
+    )
+    .reduce(
+      (acc, [importName, importPath]) => ({
+        ...acc,
+        ...(importName && importPath && { [importName]: importPath }),
+      }),
+      {},
+    );
 
   // Format entry point inputs
   const defaultEntrypoint = [
@@ -81,7 +94,7 @@ export function normalizeOptions(
     if (!partialTargets) return acc;
 
     return deepmerge(acc, partialTargets);
-  }, {} as Record<string, Partial<TargetConfiguration>>);
+  }, {} as Record<string, Partial<TargetConfiguration> | null>);
 
   // Process import path if the option is not provided
   const importPath = `@${npmScope}/${directory ? `${directory}-` : ''}${name}`;
@@ -104,6 +117,7 @@ export function normalizeOptions(
     importPath,
     hapifyModelsJsonRelativePath,
     hapifyImportReplacements,
+    importPrefixPath: getImportPrefixPath(tree, directory),
     templates,
     secondaryEntrypoints,
     targets,